pyproject.toml
[project]
name = "visual-search-engine"
version = "0.1.0"
description = "Visual Search Engine using ResNet & Triplet Loss"
readme = "README.md"
authors = [
    { name = "Paweł Litwin", email = "litwinp2006@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "pandas",
    "pillow",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-multipart",
    "requests"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

# --- RUFF (Linter & Formatter) ---
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["torchvision.*", "PIL.*", "pandas.*", "matplotlib.*", "sklearn.*"]
ignore_missing_imports = true
